+++ app/static/lib/annotation.js
@@ -1,8 +1,21 @@
-import { v, cm, log, translator } from './main.js';
+import { v, cm, log, translator, meiFileLocation, fileLocationType } from './main.js';
 import { convertCoords, generateXmlId, rmHash, setCursorToId } from './utils.js';
 import { meiNameSpace, xmlNameSpace, xmlToString } from './dom-utils.js';
-import { circle, diffRemoved, highlight, fileCode, link, pencil, rdf, symLinkFile } from '../css/icons.js';
+import { circle, diffRemoved, highlight, fileCode, identify, link, pencil, rdf, symLinkFile } from '../css/icons.js';
 import { removeInEditor } from './editor.js';
+import {
+  solid,
+  getSolidStorage,
+  friendContainer,
+  annotationContainer,
+  establishContainerResource,
+  establishResource,
+  createMAOMusicalObject,
+} from './solid.js';
+import { 
+  nsp,
+  traverseAndFetch
+} from './linked-data.js'
 
 export let annotations = [];
 
@@ -43,16 +56,16 @@ export function refreshAnnotationsList() {
     annoListItemButtons.classList.add('annotationListItemButtons');
     const flipToAnno = document.createElement('a');
     flipToAnno.insertAdjacentHTML('afterbegin', symLinkFile); //flipToEncoding;
-    flipToAnno.id = 'flipPageToAnnotationText';
+    flipToAnno.classList.add('flipPageToAnnotationText');
     flipToAnno.title = translator.lang.flipPageToAnnotationText.description;
     flipToAnno.classList.add('icon');
     if (!'selection' in a) flipToAnno.classList.add('disabled');
     const deleteAnno = document.createElement('a');
-    deleteAnno.id = 'deleteAnnotation';
+    deleteAnno.classList.add('deleteAnnotation');
     deleteAnno.insertAdjacentHTML('afterbegin', diffRemoved);
     deleteAnno.title = translator.lang.deleteAnnotation.description;
     const isStandoff = document.createElement('a');
-    isStandoff.id = 'makeStandOffAnnotation';
+    isStandoff.classList.add('makeStandOffAnnotation');
     isStandoff.insertAdjacentHTML('afterbegin', rdf);
     isStandoff.title = translator.lang.makeStandOffAnnotation.description;
     isStandoff.classList.add('icon');
@@ -63,7 +76,7 @@ export function refreshAnnotationsList() {
     } else {
       (isStandoff.title = translator.lang.makeStandOffAnnotation.descriptionToLocal + ': '), a.id;
       isStandoff.dataset.id = a.id;
-      isStandoff.addEventListener('click', copyIdToClipboard);
+      isStandoff.addEventListener('click', (e) => window.open(a.id, "_blank"));
     }
     const isInline = document.createElement('a');
     isInline.id = 'makeInlineAnnotation';
@@ -92,6 +105,10 @@ export function refreshAnnotationsList() {
       case 'annotateDescribe':
         summary.insertAdjacentHTML('afterbegin', pencil);
         details.insertAdjacentHTML('afterbegin', a.description);
+        break;
+      case 'annotateIdentify':
+        summary.insertAdjacentHTML('afterbegin', identify);
+        details.insertAdjacentHTML('afterbegin', `<div class="mao-musMat" id="musMat_${a.id}"></div>`);
         break;
       default:
         console.warn('Unknown type when drawing annotation in list: ', a);
@@ -125,20 +142,28 @@ export function refreshAnnotationsList() {
 }
 
 export function generateAnnotationLocationLabel(a) {
+  console.log("generating anno label for ", a)
   const annotationLocationLabel = document.createElement('span');
-  if (a.firstPage === 'meiHead') {
-    annotationLocationLabel.innerHTML = `MEI&nbsp;head&nbsp;(${a.selection.length}&nbsp;${translator.lang.elementsPlural.text})`;
-  } else if (a.firstPage === 'unsituated' || a.firstPage < 0) {
-    annotationLocationLabel.innerHTML = 'Unsituated';
-  } else {
-    annotationLocationLabel.innerHTML =
-      translator.lang.pageAbbreviation.text +
-      '&nbsp;' +
-      (a.firstPage === a.lastPage ? a.firstPage : a.firstPage + '&ndash;' + a.lastPage) +
-      ` (${a.selection.length}&nbsp;${translator.lang.elementsPlural.text})`;
+  if(a.type === "annotateIdentify") { 
+    // special case: identified MAO objects
+    // just add a placeholder for a label and exit
+    // (since multiple selections may be situated at different locations, we deal with them elsewhere)
+    annotationLocationLabel.innerHTML = '<span class="label"></span>';
+  } else { 
+    if (a.firstPage === 'meiHead') {
+      annotationLocationLabel.innerHTML = `MEI&nbsp;head&nbsp;(${a.selection.length}&nbsp;${translator.lang.elementsPlural.text})`;
+    } else if (a.firstPage === 'unsituated' || a.firstPage < 0) {
+      annotationLocationLabel.innerHTML = 'Unsituated';
+    } else {
+      annotationLocationLabel.innerHTML =
+        translator.lang.pageAbbreviation.text +
+        '&nbsp;' +
+        (a.firstPage === a.lastPage ? a.firstPage : a.firstPage + '&ndash;' + a.lastPage) +
+        ` (${a.selection.length}&nbsp;${translator.lang.elementsPlural.text})`;
+    }
   }
   annotationLocationLabel.classList.add('annotationLocationLabel');
-  annotationLocationLabel.dataset.id = a.id;
+  annotationLocationLabel.dataset.id = "loc-"+a.id;
   return annotationLocationLabel;
 }
 
@@ -150,17 +175,34 @@ export function situateAnnotations() {
     a.firstPage = 'unsituated';
     a.lastPage = -1;
     if ('selection' in a) {
-      a.firstPage = await v.getPageWithElement(a.selection[0]);
-      a.lastPage = await v.getPageWithElement(a.selection[a.selection.length - 1]);
+      let selectionStart = a.selection[0].indexOf("#") < 0 
+        ? a.selection[0] : a.selection[0].substr(a.selection[0].indexOf("#")+1)
+      let selLen = a.selection.length-1;
+      let selectionEnd = a.selection[selLen].indexOf("#") < 0 
+        ? a.selection[selLen] : a.selection[selLen].substr(a.selection[selLen].indexOf("#")+1)
+      a.firstPage = await v.getPageWithElement(selectionStart);
+      a.lastPage = await v.getPageWithElement(selectionEnd);
       if (a.firstPage < 0 && v.speedMode) {
         if (v.xmlDoc.querySelector('[*|id=' + a.selection[0] + ']')?.closest('meiHead')) a.firstPage = 'meiHead';
         else console.warn('Cannot locate annotation ', a);
       } else {
         // if not speed mode, asynchronous return of page numbers after we are finished here
-        const annotationLocationLabelElement = document.querySelector(`.annotationLocationLabel[data-id=${a.id}`);
+        const annotationLocationLabelElement = document.querySelector(
+          `.annotationLocationLabel[data-id=${CSS.escape(a.id)}`
+        );
         if (annotationLocationLabelElement) {
           annotationLocationLabelElement.innerHTML = generateAnnotationLocationLabel(a).innerHTML;
         }
+      }
+    }
+    // for any identifying annotations, see if there are new extracts associated with the musical material
+    if(a.type == "annotateIdentify") { 
+      // find it in the list
+      const musMatDiv = document.querySelector('#musMat_'+CSS.escape(a.id));
+      console.log("extracts DIV: ", musMatDiv);
+      if(musMatDiv && !musMatDiv.innerHTML.length) { 
+        musMatDiv.closest(".annotationListItem").querySelector(".makeStandoffAnnotation svg").classList.add("clockwise");
+        fetchMAOComponentsForIdentifiedObject(a.id);
       }
     }
   });
@@ -177,6 +219,18 @@ export function deleteAnnotation(uuid) {
 }
 
 // functions to draw annotations
+
+async function drawIdentify(a) {
+  if('selection' in a) { 
+    const els = a.selection.map((s) => document.getElementById(s));
+    els
+      .filter((e) => e !== null) // (null if not on current page)
+      .forEach((e) => e.classList.add('annotationIdentify'));
+  } else {
+    console.warn('failing to draw identify annotation without selection: ', a);
+  }
+}
+
 function drawHighlight(a) {
   if ('selection' in a) {
     const els = a.selection.map((s) => document.getElementById(s));
@@ -273,6 +327,9 @@ export function refreshAnnotations(forceListRefresh = false) {
     annotations.forEach((a) => {
       if ('type' in a) {
         switch (a.type) {
+          case 'annotateIdentify':
+            drawIdentify(a);
+            break;
           case 'annotateHighlight':
             drawHighlight(a);
             break;
@@ -300,9 +357,12 @@ export function addAnnotationHandlers() {
   // TODO extend this to allow app to consume (TROMPA-style) Annotation Toolkit descriptions
 
   const annotationHandler = (e) => {
-    console.log('Clicked to make new annotation!', e);
-    console.log('Selected elements: ', v.selectedElements);
-    switch (e.target.getAttribute('id')) {
+    console.log('annotation Handler: Clicked to make new annotation!', e);
+    console.log('annotation Handler: Selected elements: ', v.selectedElements);
+    switch (e.target.closest('.annotationToolsIcon')?.getAttribute('id')) {
+      case 'annotateIdentify':
+        createIdentify(e);
+        break;
       case 'annotateHighlight':
         createHighlight(e);
         break;
@@ -322,6 +382,32 @@ export function addAnnotationHandlers() {
   };
 
   // functions to create annotations
+  const createIdentify = (e) => {
+    const label = window.prompt("Add label for identified object (optional)");
+    const selection = v.selectedElements;
+    document.getElementById('solid_logo').classList.add('clockwise');
+    createMAOMusicalObject(selection, label)
+      .then((maoMusicalMaterial) => {
+        getSolidStorage().then((solidStorage) => {
+          console.log(
+            'CREATED MUSICAL MATERIAL: ',
+            solidStorage + maoMusicalMaterial.headers.get('location').substr(1), maoMusicalMaterial
+          );
+          const a = {
+            id: solidStorage + maoMusicalMaterial.headers.get('location').substr(1),
+            type: 'annotateIdentify',
+            selection: selection,
+            isStandoff: true,
+          };
+          annotations.push(a);
+          refreshAnnotations(true);
+        });
+      })
+      .finally(() => {
+        document.getElementById('solid_logo').classList.remove('clockwise');
+      });
+    // writing inline not supported
+  };
   const createHighlight = (e) => {
     const a = {
       id: generateXmlId('annot', v.xmlIdStyle),
@@ -330,6 +416,7 @@ export function addAnnotationHandlers() {
     };
     annotations.push(a);
     writeInlineIfRequested(a);
+    writeStandoffIfRequested(a);
   };
   const createCircle = (e) => {
     const a = {
@@ -339,6 +426,7 @@ export function addAnnotationHandlers() {
     };
     annotations.push(a);
     writeInlineIfRequested(a);
+    writeStandoffIfRequested(a);
   };
   const createDescribe = (e) => {
     // TODO improve UX!
@@ -351,6 +439,7 @@ export function addAnnotationHandlers() {
     };
     annotations.push(a);
     writeInlineIfRequested(a);
+    writeStandoffIfRequested(a);
   };
   const createLink = (e) => {
     // TODO improve UX!
@@ -364,10 +453,25 @@ export function addAnnotationHandlers() {
     };
     annotations.push(a);
     writeInlineIfRequested(a);
+    writeStandoffIfRequested(a);
   };
 
   document.querySelectorAll('.annotationToolsIcon').forEach((a) => a.removeEventListener('click', annotationHandler));
   document.querySelectorAll('.annotationToolsIcon').forEach((a) => a.addEventListener('click', annotationHandler));
+
+  // disable 'identify music object' unless 'linked data' domain selected
+  document.querySelectorAll('.annotationToolsDomainSelectionItem input').forEach((i) => i.removeEventListener('click', enableDisableIdentifyObject))
+  document.querySelectorAll('.annotationToolsDomainSelectionItem input').forEach((i) => i.addEventListener('click', enableDisableIdentifyObject))
+  enableDisableIdentifyObject(); // set initial status
+}
+
+function enableDisableIdentifyObject() { 
+  let identifyTool = document.getElementById("annotateIdentify");
+  if(document.getElementById("writeAnnotationInline").checked) { 
+    identifyTool.classList.add("disabled");
+  } else { 
+    identifyTool.classList.remove("disabled");
+  }
 }
 
 // reads <annot> elements from XML DOM and adds them into annotations array
@@ -395,9 +499,7 @@ export function readAnnots(flagLimit = false) {
     if (annot.textContent) {
       const ptrs = annot.getElementsByTagNameNS(meiNameSpace, 'ptr');
       if (ptrs.length) {
-        console.log(ptrs);
         annotation.type = 'annotateLink';
-        console.log('ptrs: ', ptrs);
         annotation.url = ptrs[0].getAttribute('target');
       } else {
         annotation.type = 'annotateDescribe';
@@ -502,6 +604,7 @@ export function deleteAnnot(xmlId) {
   }
 }
 
+// GUI function to set up 'Load linked data' UI
 export function loadWebAnnotation(prev = '') {
   let msg = translator.lang.loadWebAnnotationMessage.text;
   let prevurl = '';
@@ -525,8 +628,48 @@ export function loadWebAnnotation(prev = '') {
   let urlstr = window.prompt(msg, prevurl);
   if (urlstr) {
     if (!(urlstr.startsWith('http://') || urlstr.startsWith('https://'))) urlstr = 'https://' + urlstr;
-    fetchWebAnnotations(urlstr);
+    try {
+      // ensure working URLs provided
+      attemptFetchExternalResource(
+        new URL(urlstr), // traversal start
+        [new URL(nsp.OA + 'Annotation')], // target types
+        {
+          typeToHandlerMap: {
+            [nsp.OA + 'Annotation']: {
+              func: ingestWebAnnotation
+            }
+          },
+          followList: [new URL(nsp.LDP + 'contains')], // predicates to traverse
+          fetchMethod: solid.getDefaultSession().info.isLoggedIn ? solid.fetch : fetch
+        }
+      );
+    } catch (e) {
+      // invalid URL
+      loadWebAnnotation();
+    }
   }
+}
+
+// Wrapper around traverseAndFetch that reports back errors / progress to 'Load linked data' UI
+export function attemptFetchExternalResource(url, targetTypes, configObj) { 
+  console.log('fetch external resource: ', url, targetTypes, typeToHandlerMap, followList, blockList, jumps);
+  // spin the icon to indicate loading activity
+  const icon = document.getElementById('addWebAnnotationIcon');
+  const svgs = Array.from(icon.getElementsByTagName('svg'));
+  svgs.forEach((t) => t.classList.add('clockwise'));
+  traverseAndFetch(url, targetTypes, configObj)
+    .catch((resp) => {
+        if (userProvided) {
+          console.error("Couldn't load external resource, error response: ", resp);
+          // user-provided url didn't work, so hand control back to user
+          if (resp.url) loadWebAnnotation(resp);
+          else loadWebAnnotation(url.href);
+        }
+      })
+    .finally(() => {
+      // notify that we've stopped loading
+      svgs.forEach((t) => t.classList.remove('clockwise'));
+    });
 }
 
 export function fetchWebAnnotations(url, userProvided = true, jumps = 10) {
@@ -534,7 +677,8 @@ export function fetchWebAnnotations(url, userProvided = true, jumps = 10) {
   const icon = document.getElementById('addWebAnnotationIcon');
   const svgs = Array.from(icon.getElementsByTagName('svg'));
   svgs.forEach((t) => t.classList.add('clockwise'));
-  fetch(url, {
+  let fetchAppropriately = solid.getDefaultSession().info.isLoggedIn ? solid.fetch : fetch;
+  fetchAppropriately(url, {
     headers: {
       Accept: 'application/ld+json',
     },
@@ -546,8 +690,8 @@ export function fetchWebAnnotations(url, userProvided = true, jumps = 10) {
         return resp.json();
       }
     })
+    .then((json) => jsonld.expand(json))
     .then((json) => {
-      console.log('json response: ', json);
       let resourceDescription;
       if (Array.isArray(json)) {
         resourceDescription = json.find((o) => o['@id'] === url);
@@ -559,7 +703,6 @@ export function fetchWebAnnotations(url, userProvided = true, jumps = 10) {
         resourceDescription = json;
       }
       if (resourceDescription && '@type' in resourceDescription) {
-        console.log('found resource desc: ', resourceDescription);
         if (resourceDescription['@type'].includes('http://www.w3.org/ns/ldp#Container')) {
           // found a container, recurse on members
           if ('http://www.w3.org/ns/ldp#contains' in resourceDescription) {
@@ -576,7 +719,6 @@ export function fetchWebAnnotations(url, userProvided = true, jumps = 10) {
           }
         } else if (resourceDescription['@type'].includes('http://www.w3.org/ns/oa#Annotation')) {
           // found an annotation!
-          console.log('Found annotation!!', resourceDescription);
           ingestWebAnnotation(resourceDescription);
           return resourceDescription;
         } else {
@@ -623,29 +765,23 @@ export function ingestWebAnnotation(webAnno) {
           // decare a linking annotation
           anno.url = firstBody['@id'];
           anno.type = 'annotateLink';
-        } else if (
-          '@type' in firstBody &&
-          'http://www.w3.org/1999/02/22-rdf-syntax-ns#value' in firstBody &&
-          'http://www.w3.org/ns/oa#TextualBody' in firstBody['@type']
-        ) {
+        } else if ('@type' in firstBody && nsp.RDF + value in firstBody && nsp.OA + 'TextualBody' in firstBody['@type']) {
           // declare a describing annotation
           console.log('Declaring a describing annotation!');
-          anno.description = firstBody['http://www.w3.org/1999/02/22-rdf-syntax-ns#'];
+          anno.description = firstBody[nsp.RDF + value];
           anno.type = 'annotateDescribe';
         } else {
           console.log("Don't know how to handle body of this annotation: ", anno);
         }
       }
-    } else if ('http://www.w3.org/ns/oa#bodyValue' in webAnno) {
+    } else if (nsp.OA + 'bodyValue' in webAnno) {
       // declare describing annotation
       console.log('Declaring a describing annotation!');
       // TODO decide what to do for multiple bodies
-      anno.description = webAnno['http://www.w3.org/ns/oa#bodyValue'][0]['@value'];
+      anno.description = webAnno[nsp.OA + 'bodyValue'][0]['@value'];
       anno.type = 'annotateDescribe';
     }
 
-    console.log('Annotations: ', annotations);
-    console.log('Anno: ', anno);
     anno.selection = targets.map((t) => t['@id'].split('#')[1]);
     if (annotations.findIndex((a) => a.id === anno.id) < 0) {
       // add to list if we don't already have it
@@ -658,7 +794,7 @@ export function ingestWebAnnotation(webAnno) {
 
 function writeInlineIfRequested(a) {
   // write annotation to inline <annot> if the user has requested this
-  if (document.getElementById('writeAnnotInline').checked) {
+  if (document.getElementById('writeAnnotationInline').checked) {
     let el = v.xmlDoc.querySelector('[*|id="' + v.selectedElements[0] + '"]');
     if (el) {
       let payload;
@@ -671,6 +807,223 @@ function writeInlineIfRequested(a) {
       writeAnnot(el, a.id, a.selection, payload);
       a.isInline = true;
     } else console.warn('writeInlineIfRequested: Cannot find beforeThis element for ' + a.id);
+  }
+}
+
+
+// fetch all components hanging off a musical material
+async function fetchMAOComponentsForIdentifiedObject(musMatUrl) {
+  traverseAndFetch(
+    new URL(musMatUrl), 
+    [new URL(nsp.MAO + "MusicalMaterial"), new URL(nsp.MAO + "Extract"), new URL(nsp.MAO + "Selection")],
+    { 
+      typeToHandlerMap: { 
+        [nsp.MAO + "MusicalMaterial"]: { 
+          func: drawMusicalMaterialForIdentifiedObject
+        }, 
+        [nsp.MAO + "Extract"]: {
+          func: drawExtractsForIdentifiedObject
+        }, 
+        [nsp.MAO + "Selection"]: { 
+          func: drawSelectionsForIdentifiedObject
+        }
+      },
+      followList: [new URL(nsp.FRBR + "embodiment"), new URL(nsp.MAO + "setting")],
+      fetchMethod: solid.getDefaultSession().info.isLoggedIn ? solid.fetch : fetch
+    }
+  ).catch(e => { 
+    log("Couldn't load extracts associated with identified musical object:", e);
+    const loadingIndicator = document.querySelector('#musMat_'+CSS.escape(musMatUrl)+' span');
+    if(loadingIndicator) { 
+      loadingIndicator.classList.remove("clockwise");
+    }
+  })
+}
+
+/*
+async function fetchExtractsForIdentifiedObject(url) { 
+  traverseAndFetch(
+    new URL(url), 
+    [new URL(nsp.MAO + "MusicalMaterial")],
+    { 
+      typeToHandlerMap: { 
+        [nsp.MAO + "MusicalMaterial"]: { 
+          func: drawExtractsForIdentifiedObject,
+          args: [url]
+        }
+      },
+      fetchMethod: solid.getDefaultSession().info.isLoggedIn ? solid.fetch : fetch
+    }
+  ).catch(e => { 
+    log("Couldn't load extracts associated with identified musical object:", e);
+    const loadingIndicator = document.querySelector('#musMat_'+CSS.escape(url)+' span');
+    if(loadingIndicator) { 
+      loadingIndicator.classList.remove("clockwise");
+    }
+  })
+}
+*/
+
+/*
+* Draw placeholder for MAO setting into the musMat element corresponding to this url
+* Check first whether this setting is already included. Do not overwrite other settings.
+*/
+async function drawMusicalMaterialForIdentifiedObject(obj, url){
+  console.log("drawMusMatForIdentifiedObject: ", obj, url);
+  const musMat = document.getElementById("musMat_"+url.href);
+  if(musMat) { 
+    // if we have a label and haven't drawn one already...
+    const myListItem = musMat.closest(".annotationListItem");
+    const myLabel = myListItem.querySelector(".annotationLocationLabel")
+    console.log("LABEL: ", obj, myLabel)
+    if(nsp.RDFS+"label" in obj && !myLabel.innerText.length) { 
+      const label = Array.isArray(obj[nsp.RDFS+"label"]) ? 
+        obj[nsp.RDFS+"label"] : [obj[RDFS+"label"]];
+      myLabel.innerText = label[0]["@value"]; // TODO support multiple labels?
+    }
+    if(nsp.MAO+"setting" in obj) {
+      const alreadyIncluded = musMat.querySelectorAll(".mao-extract");
+      const alreadyIncludedUrls = alreadyIncluded.forEach(n => n.id.replace("extract", ""));
+      let extractsToAdd;
+      if(alreadyIncludedUrls) { 
+        extractsToAdd = obj[nsp.MAO+"setting"].filter(s => !alreadyIncludedUrls.includes(s["@id"]));
+      } else {
+        extractsToAdd = obj[nsp.MAO+"setting"];
+      }
+      console.debug("Trying to add extracts... ", extractsToAdd)
+      extractsToAdd.forEach(s => {
+        musMat.insertAdjacentHTML('beforeend', `<div class="mao-extract" id="extract_${s["@id"]}"></div>`);
+      })
+      console.debug("Done with musMat: ", musMat)
+    } else { 
+      console.warn("Can't draw a MusicalMaterial without a setting:", obj)
+    }
+  } else { 
+    console.warn("Trying to draw musical material for identified object but have no musMat div to hook it into", obj, url);
+  }
+}
+
+async function drawExtractsForIdentifiedObject(obj, url) { 
+  console.log("drawExtractsForIdentifiedObject: ", obj, url);
+  const extract = document.getElementById("extract_"+url.href);
+  if(extract) { 
+    if(nsp.FRBR + "embodiment" in obj) {
+      const alreadyIncluded = extract.querySelectorAll(".mao-selection");
+      const alreadyIncludedUrls = alreadyIncluded.forEach(n => n.id.replace("selection_", ""));
+      let selectionsToAdd;
+      if(alreadyIncludedUrls) { 
+        selectionsToAdd = obj[nsp.FRBR+"embodiment"].filter(s => !alreadyIncludedUrls.includes(s["@id"]));
+      } else { 
+        selectionsToAdd = obj[nsp.FRBR+"embodiment"];
+      }
+      selectionsToAdd.forEach(s => {
+        extract.insertAdjacentHTML('beforeend', `<div class="mao-selection" id="selection_${s["@id"]}"></div>`)
+      });
+    } else { 
+      console.warn("Can't draw an Extract without an embodiment:", obj)
+    }
+  } else { 
+    console.warn("Trying to draw extract for identified object but have no extract div to hook it into", obj, url);
+  }
+}
+
+async function drawSelectionsForIdentifiedObject(obj, url) { 
+  console.log("drawSelectionForIdentifiedObject: ", obj, url);
+  const selection = document.getElementById("selection_"+url.href);
+  if(selection){ 
+    if(nsp.FRBR+"part" in obj) {
+      try { 
+        const a = {};
+        a.selection = obj[nsp.FRBR+"part"].map(
+          s => s["@id"].substr(s["@id"].lastIndexOf("#")+1)
+        )
+        a.firstPage = await v.getPageWithElement(a.selection[0]);
+        a.lastPage = await v.getPageWithElement(a.selection[a.selection.length-1]);
+        selection.innerText = generateAnnotationLocationLabel(a).innerText;
+      }
+      catch(e) { 
+        console.warn("Couldn't situate FRBR:parts of selection: ", e);
+      }
+    } else { 
+      console.warn("Can't draw a selection without parts: ", obj, url);
+    }
+    const myMusMat = selection.closest(".annotationListItem");
+    if(myMusMat) { 
+      const myLoadingIndicator = myMusMat.querySelector(".makeStandoffAnnotation svg");
+      if(myLoadingIndicator)
+        myLoadingIndicator.classList.remove("clockwise");
+    }
+  }else { 
+    console.warn("Trying to draw selection for identified object but have no selection div to hook it into", obj, url);
+  }
+}
+
+
+
+async function writeStandoffIfRequested(a) {
+  // write to a stand-off Web Annotation in the user's Solid Pod if requested
+  if (document.getElementById('writeAnnotationStandoff').checked) {
+    a.isStandoff = true;
+    if (solid.getDefaultSession().info.isLoggedIn) {
+      document.getElementById('solid_logo').classList.add('clockwise');
+      // ensure mei-friend container exists
+      establishContainerResource(friendContainer)
+        .then((friendContainerResource) => {
+          establishContainerResource(annotationContainer)
+            .then((annotationContainerResource) => {
+              // generate a web annotation JSON-LD object
+              let webAnno = new Object();
+              let body = new Object();
+              webAnno['@type'] = [nsp.OA + 'Annotation'];
+              webAnno[nsp.OA + 'hasTarget'] = a.selection.map((s) => {
+                // TODO: do something clever if fileLocationType = "file" (local)
+                return { '@id': meiFileLocation + '#' + s };
+              });
+              switch (a.type) {
+                case 'annotateHighlight':
+                  webAnno[nsp.OA + 'motivatedBy'] = [{ '@id': nsp.OA + 'highlighting' }];
+                  break;
+                case 'annotateDescribe':
+                  body['@type'] = [nsp.OA + 'TextualBody'];
+                  if ('description' in a) {
+                    body[nsp.RDF + 'value'] = a.description;
+                  } else {
+                    console.warn('Describing annotation without a description: ', a);
+                  }
+                  webAnno[nsp.OA + 'motivatedBy'] = [{ '@id': nsp.OA + 'describing' }];
+                  webAnno[nsp.OA + 'hasBody'] = [body];
+                  break;
+                case 'annotateLink':
+                  webAnno[nsp.OA + 'motivatedBy'] = [{ '@id': nsp.OA + 'linking' }];
+                  webAnno[nsp.OA + 'hasBody'] = [{ '@id': a.url }];
+                  break;
+                default:
+                  console.warn('Trying to write standoff annotation with unknown annotation type:', a);
+                  break;
+              }
+              webAnno['@id'] = annotationContainerResource + a.id;
+              establishResource(webAnno['@id'], webAnno).then((webAnnoResp) => {
+                if (webAnnoResp.ok) {
+                  console.log('Success! Posted Web Annotation:', webAnno);
+                } else {
+                  console.warn("Couldn't post WebAnno: ", webAnno, webAnnoResp);
+                  throw webAnnoResp;
+                }
+              });
+            })
+            .catch((e) => {
+              console.error("Couldn't post WebAnno:", e);
+            });
+        })
+        .catch(() => {
+          console.error("Couldn't establish container:", friendContainer);
+        })
+        .finally(() => {
+          document.getElementById('solid_logo').classList.remove('clockwise');
+        });
+    } else {
+      log('Cannot write standoff annotation: Please ensure you are logged in to a Solid Pod');
+    }
   }
 }