+++ app/static/lib/storage.js
@@ -52,6 +52,7 @@ export default class Storage {
       this._breaks = this.storage.getItem('breaks');
       this._forkAndOpen = this.storage.getItem('forkAndOpen');
       this._githubLogoutRequested = this.storage.getItem('githubLogoutRequested');
+      this._restoreSolidSession = this.storage.getItem('restoreSolidSession');
       //fileChangedFromStorage = fileChangedFromStorage ? parseInt(storage.getItem("fileChanged")) : 0;
     }
   }
@@ -60,6 +61,16 @@ export default class Storage {
     if (this.supported) {
       this.storage.clear();
     }
+  }
+
+  clearSolid() { 
+    // remove data related to Solid credentials from local storage after successful login
+    let keys = Object.keys(this.storage);
+    keys.forEach(k => { 
+      if(k.startsWith("solidClient") || k.startsWith("issuerConfig")) {
+        this.storage.removeItem(k);
+      }
+    })
   }
 
   removeItem(item) {
@@ -299,4 +310,14 @@ export default class Storage {
     this.safelySetStorageItem('forkAndOpen', forkAndOpen);
     this._forkAndOpen = forkAndOpen;
   }
+
+  get restoreSolidSession() {
+    return this._restoreSolidSession;
+  }
+
+  set restoreSolidSession(restoreSolidSession) {
+    this.safelySetStorageItem('restoreSolidSession', restoreSolidSession);
+    this._restoreSolidSession = restoreSolidSession;
+  }
+
 }